added logout button
