+++ package.json
@@ -10,6 +10,7 @@
     "react-dom": "^16.3.2",
     "react-parallax": "^1.7.0",
     "react-redux": "^5.0.7",
+    "react-router": "^4.2.0",
     "react-router-dom": "^4.2.2",
     "react-router-redux": "^5.0.0-alpha.9",
     "react-scripts": "1.1.4",
