+++ src/action-types/user.js
@@ -1,6 +1,7 @@
 const UserTypes = {
   LOG_IN: 'LOG_IN',
   GET_USER: 'GET_USER',
+  LOG_OUT: 'LOG_OUT',
 };
 
 export default UserTypes;
